refactor: Move env parsing of deployment target to rustc_session
